requirements.txt
transformers==4.31.0
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
flask==2.3.3
pyyaml==6.0.1
dataclasses==0.6
sentence-transformers==2.2.2
